Aggiunta la funzionalita' di filtro. Allo stato dell'arte, il progetto ha le funzionalita' minime richieste. Adesso si può passare alla fase di revisione del progetto, quindi: const correttezza, gestione della memoria e altre funzionalità, come la scelta delle immagini per ciascun prodotto al momento della creazione di esso o della sua modifica.

+++ service/PriceMatcher.cpp
@@ -2,7 +2,7 @@
 
 
 
-PriceMatcher::PriceMatcher(const unsigned int lowerBound, const unsigned int upperBound) : lowerBound(lowerBound), upperBound(upperBound) { }
+PriceMatcher::PriceMatcher(const float lowerBound, const float upperBound) : lowerBound(lowerBound), upperBound(upperBound) { }
 
 bool PriceMatcher::hasMatch(const AbstractProduct& product) const {
     return (product.getPrezzo() > lowerBound && product.getPrezzo() < upperBound);
